Add packages to make code more OOP and readable
Add packages Command,Main,Exceptions,Tasks and Parser, More will be added later
