+++ src/main/java/Exceptions/EmptyDeadlineException.java
@@ -1,3 +1,5 @@
+package Exceptions;
+
 public class EmptyDeadlineException extends Exception{
 
     public EmptyDeadlineException(String desc) {
