+++ src/main/java/Exceptions/EmptyDescException.java
@@ -1,3 +1,5 @@
+package Exceptions;
+
 public class EmptyDescException extends Exception{
 
     public EmptyDescException(String desc) {
