+++ src/main/java/Exceptions/EmptyEventException.java
@@ -1,3 +1,5 @@
+package Exceptions;
+
 public class EmptyEventException extends Exception{
 
     public EmptyEventException(String desc) {
