+++ src/main/java/Exceptions/EmptyEventTimingException.java
@@ -1,3 +1,5 @@
+package Exceptions;
+
 public class EmptyEventTimingException extends Exception{
 
     public EmptyEventTimingException(String message) {
