+++ src/main/java/Exceptions/UnknownCommandException.java
@@ -1,3 +1,5 @@
+package Exceptions;
+
 public class UnknownCommandException extends Exception{
 
     public UnknownCommandException(String message) {
