+++ src/main/java/Main/FileUI.java
@@ -1,3 +1,5 @@
+package Main;
+
 import java.io.File;
 import java.io.IOException;
