+++ src/main/java/Parser/Parser.java
@@ -1,4 +1,6 @@
-import java.util.Scanner;
+package Parser;
+
+import Commands.*;
 
 public class Parser {
