+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
     __________________________________________
-    Hey! I'm Flash
+    Hey! I'm Main.Flash
     What can I do for ya?
     __________________________________________
     __________________________________________
